Fix for topology crash on load when application group is selected
